create a seperate component for demo login button
