+++ client/src/pages/Login/useStyles.ts
@@ -14,9 +14,6 @@ const useStyles = makeStyles((theme) => ({
     flexDirection: 'column',
     minHeight: '100vh',
     paddingTop: 23,
-    '& .makeStyles-authHeader-6': {
-      justifyContent: 'center',
-    },
   },
   welcome: {
     fontSize: 26,
