+++ client/src/pages/SignUp/SignUpForm/SignUpForm.tsx
@@ -6,6 +6,7 @@ import * as Yup from 'yup';
 import Typography from '@material-ui/core/Typography';
 import useStyles from './useStyles';
 import { CircularProgress } from '@material-ui/core';
+import DemoLoginButton from '../../../components/DemoLoginButton/DemoLoginButton';
 
 interface Props {
   handleSubmit: (
@@ -27,10 +28,9 @@ interface Props {
       username: string;
     }>,
   ) => void;
-  demoLogin: ({}) => void;
 }
 
-const SignUpForm = ({ handleSubmit, demoLogin }: Props): JSX.Element => {
+const SignUpForm = ({ handleSubmit }: Props): JSX.Element => {
   const classes = useStyles();
 
   return (
@@ -112,16 +112,7 @@ const SignUpForm = ({ handleSubmit, demoLogin }: Props): JSX.Element => {
             <Button type="submit" size="large" variant="contained" color="primary" className={classes.submit}>
               {isSubmitting ? <CircularProgress style={{ color: 'white' }} /> : 'Create'}
             </Button>
-            <Button
-              onClick={demoLogin}
-              type="button"
-              size="large"
-              variant="contained"
-              color="secondary"
-              className={classes.submit}
-            >
-              Login as Demo
-            </Button>
+            <DemoLoginButton />
           </Box>
         </form>
       )}
